Add files via upload
This is the google colab notebook. Downloading this notebook will defeat the purpose of using it as I am using it for the gpu resources. It has yet to be automated as I haven't figured out a way of connecting it to the pycharm file, but once I figure out how or if I finde a better way I will update this code